MVP actually done now. All features working.
