+++ src/components/AddedFeature.js
@@ -1,4 +1,6 @@
 import React from 'react';
+import { removeFeature } from "../actions/removeFeatureAction";
+import { connect } from "react-redux";
 
 const AddedFeature = props => {
   return (
@@ -6,7 +8,10 @@ const AddedFeature = props => {
       {/* Add an onClick to run a function to remove a feature */}
       <button 
       className="button"
-      onClick={() => props.removeFeature(props.feature)}
+      onClick={(e) => {
+        e.preventDefault()
+        props.removeFeature(props.feature)
+      }}
       >
         X
         </button>
@@ -15,4 +20,4 @@ const AddedFeature = props => {
   );
 };
 
-export default AddedFeature;
+export default connect(null, { removeFeature })(AddedFeature);
